Use separate request handler script + run async

+++ lab4/client.sh
@@ -1,20 +1,22 @@
 #!/bin/bash
-
 SERVER_ADDRESS=$1
 ARGUMENT=$2
-
 CLIENT_FIFO="$HOME/client_fifo"
-SERVER_FIFO="$HOME/server_fifo"
+SERVER_FIFO="/home/$SERVER_ADDRESS/server_fifo"
 
-rm $CLIENT_FIFO
+if [ -p $CLIENT_FIFO ]; then
+    rm $CLIENT_FIFO
+fi
 mkfifo $CLIENT_FIFO
 chmod 777 $CLIENT_FIFO
 
-echo $HOME > $SERVER_FIFO
-echo $ARGUMENT > $SERVER_FIFO
+echo "[CLIENT]: sending request..."
+echo $CLIENT_FIFO $ARGUMENT > $SERVER_FIFO
 
-while [[ $RESPONSE -le 0 ]]; do
+echo "[CLIENT]: waiting for response.."
+while [ -z $RESPONSE ]; do
   read RESPONSE < $CLIENT_FIFO
-  echo $RESPONSE
 done
+echo $RESPONSE
+rm $CLIENT_FIFO
 exit 0
